requirements.txt
aniso8601==7.0.0
asgiref==3.6.0
autopep8==2.0.2
Django==3.2.18
django-cors-headers==3.14.0
django-filter==23.1
django-graphql-auth==0.3.16
django-graphql-jwt==0.3.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
djongo==1.3.6
dnspython==2.3.0
graphene==2.1.9
graphene-django==2.15.0
graphql-core==2.3.2
graphql-relay==2.0.1
promise==2.3
pycodestyle==2.10.0
PyJWT==1.7.1
pymongo==4.3.3
pytz==2023.3
pywatchman==1.4.1
Rx==1.6.3
singledispatch==4.0.0
six==1.16.0
sqlparse==0.2.4
text-unidecode==1.3
tk==0.1.0


